[procmgrd] Merge ddot tests into e2e.rs, remove blanket allow(dead_code)
Consolidate ddot.rs into e2e.rs so there is a single test binary and
all helpers are used. Remove #![allow(dead_code)] from helpers. Update
BUILD.bazel to a single e2e_test target with DDOT_TEMPLATE_PATH.
Restore env var doc comment on DaemonHandle::start per PR review.
