+++ pkg/procmgr/rust/BUILD.bazel
@@ -139,12 +139,12 @@ rust_test(
 )
 
 rust_test(
-    name = "integration_test",
+    name = "e2e_test",
     srcs = [
-        "tests/ddot.rs",
+        "tests/e2e.rs",
         "tests/helpers/mod.rs",
     ],
-    crate_root = "tests/ddot.rs",
+    crate_root = "tests/e2e.rs",
     data = [
         ":dd-procmgr",
         ":dd-procmgrd",
@@ -155,29 +155,6 @@ rust_test(
         "CARGO_BIN_EXE_dd-procmgr": "$(rootpath :dd-procmgr)",
         "CARGO_BIN_EXE_dd-procmgrd": "$(rootpath :dd-procmgrd)",
         "DDOT_TEMPLATE_PATH": "$(rootpath //pkg/fleet/installer/packages/embedded/tmpl:datadog-agent-ddot.yaml.tmpl)",
-    },
-    deps = [
-        "@crates//:nix",
-        "@crates//:serde_json",
-        "@crates//:tempfile",
-    ],
-)
-
-rust_test(
-    name = "e2e_test",
-    srcs = [
-        "tests/e2e.rs",
-        "tests/helpers/mod.rs",
-    ],
-    crate_root = "tests/e2e.rs",
-    data = [
-        ":dd-procmgr",
-        ":dd-procmgrd",
-    ],
-    edition = "2024",
-    rustc_env = {
-        "CARGO_BIN_EXE_dd-procmgr": "$(rootpath :dd-procmgr)",
-        "CARGO_BIN_EXE_dd-procmgrd": "$(rootpath :dd-procmgrd)",
     },
     deps = [
         "@crates//:nix",
